+++ views/pdfLayouts/DBJR.php
@@ -167,6 +167,15 @@ class DBJR extends IPDFLayout
         $pdf->setImageScale(PDF_IMAGE_SCALE_RATIO);
         $pdf->SetFont('dejavusans', '', 10);
 
+        $pdf->setHtmlVSpace([
+            'ul'         => [['h' => 0, 'n' => 0], ['h' => 0, 'n' => 0]],
+            'li'         => [['h' => 0, 'n' => 0], ['h' => 0, 'n' => 0]],
+            'ol'         => [['h' => 0, 'n' => 0], ['h' => 0, 'n' => 0]],
+            'div'        => [['h' => 0, 'n' => 0], ['h' => 0, 'n' => 0]],
+            'p'          => [['h' => 0, 'n' => 0], ['h' => 0, 'n' => 0]],
+            'blockquote' => [['h' => 0, 'n' => 0], ['h' => 0, 'n' => 0]],
+        ]);
+
         $this->pdf = $pdf;
 
         return $pdf;
@@ -178,10 +187,9 @@ class DBJR extends IPDFLayout
      */
     public function isSkippingSectionTitles(IMotionSection $section)
     {
-        if ($section->getSettings()->title == \Yii::t('motion', 'motion_text')) {
+        if ($section->getSettings()->title === \Yii::t('motion', 'motion_text')) {
             return true;
         }
         return false;
-
     }
 }